+++ client/src/utils/upload-payload.js
@@ -1,25 +1,6 @@
-export const DEFAULT_FILE_NAME = "New File";
-export const USER_FILE_PREFIX = "gxuserfiles://";
-export const URI_PREFIXES = [
-    "http://",
-    "https://",
-    "ftp://",
-    "file://",
-    "gxfiles://",
-    "gximport://",
-    "gxuserimport://",
-    USER_FILE_PREFIX,
-    "gxftp://",
-    "drs://",
-    "invenio://",
-    "zenodo://",
-    "dataverse://",
-    "elabftw://",
-];
+import { isUrl } from "./url";
 
-export function isUrl(content) {
-    return URI_PREFIXES.some((prefix) => content.startsWith(prefix));
-}
+export const DEFAULT_FILE_NAME = "New File";
 
 export function uploadPayload(items, historyId, composite = false) {
     const files = [];
